refactor : Remove package main
